+++ favorites/reducers/products.js
@@ -10,6 +10,13 @@ import {
   REQUEST_FAVORITES,
   ERROR_FETCH_FAVORITES,
   FAVORITES_LIFETIME,
+  REQUEST_ADD_FAVORITES,
+  RECEIVE_ADD_FAVORITES,
+  ERROR_ADD_FAVORITES,
+  REQUEST_REMOVE_FAVORITES,
+  RECEIVE_REMOVE_FAVORITES,
+  ERROR_REMOVE_FAVORITES,
+  ABORT_ADD_FAVORITES,
 } from '../constants';
 
 /**
@@ -28,6 +35,27 @@ const products = (state = {}, action) => {
         expires: 0,
         // No ready here! It should be undefined or true!
       };
+    case REQUEST_ADD_FAVORITES:
+      return {
+        ...state,
+        isFetching: true,
+        ids: [
+          ...(state.ids || []),
+          action.productId,
+        ],
+      };
+    case REQUEST_REMOVE_FAVORITES:
+      return {
+        ...state,
+        isFetching: true,
+        ids: state.ids.filter(id => id !== action.productId),
+      };
+    case ABORT_ADD_FAVORITES:
+      return {
+        ...state,
+        isFetching: false,
+        ids: state.ids.filter(id => id !== action.productId),
+      };
     case RECEIVE_FAVORITES:
       return {
         ...state,
@@ -35,6 +63,27 @@ const products = (state = {}, action) => {
         expires: Date.now() + FAVORITES_LIFETIME,
         ids: action.products.map(product => product.id),
         ready: true,
+      };
+    case RECEIVE_ADD_FAVORITES:
+    case RECEIVE_REMOVE_FAVORITES:
+      return {
+        ...state,
+        isFetching: false,
+      };
+    case ERROR_ADD_FAVORITES:
+      return {
+        ...state,
+        isFetching: false,
+        ids: state.ids.filter(id => id !== action.productId),
+      };
+    case ERROR_REMOVE_FAVORITES:
+      return {
+        ...state,
+        isFetching: false,
+        ids: [
+          ...state.ids,
+          action.productId,
+        ],
       };
     case ERROR_FETCH_FAVORITES:
       return {